centralize tracking code
